pyproject.toml
[project]
name = "lfm-mcp-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp>=1.26.0",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
